Add files via upload
# First class exercise
This is an introduction to jaclang in VS Code. The first step was to integrate ubuntu on to my Windows 11 PC and proceeded to use it on VS code. A simple jac file was created basics of jac introduced. This also shows how to integrate jaclang in pure python files.